Create README.md
n recent years online shopping has gained a huge boom. With this increase, most of the features of online shopping are developed but some features like negotiating with shopkeepers are not available which is sometimes possible in offline purchasing. We have implemented a chatbot for negotiating on the products. The chatbot interacts with customers and assists them to get a satisfactory price on product(s). With such a system, which impacts on major areas of online shopping there are possibilities in which either the seller of the product or customer’s budget gets compromised. To avoid such situations we have developed an algorithm which works along with prediction of old available data to provide a price. Price prediction has less accuracy at times because either irrelevant features/attributes of data are used or some algorithms are not suitable for a particular dataset. Due to this, Ecommerce business does not directly rely on price prediction systems since even a wrong prediction of a single product can result in business losses. Some models also fail when data scales or some feature is unavailable after time on which model prediction was dependent. Then those changes are to be managed to maintain the accuracy and reliability of the model. In our chatbot system we have tried to resolve some of such issues.